+++ native/dispatch.h
@@ -13,6 +13,8 @@
 #ifndef DISPATCH_H
 #define DISPATCH_H
 
+#define MSG_SIZE 1024
+
 #include "ffi.h"
 #include "com_sun_jna_Function.h"
 #include "com_sun_jna_Native.h"
@@ -130,18 +132,23 @@ typedef struct _callback {
 #endif
 
 #if defined(_MSC_VER)
+// snprintf on windows is broken; always nul-terminate manually
+// DO NOT rely on the return value...
+static int snprintf(char * str, size_t size, const char * format, ...) {
+  int retval;
+  va_list ap;
+  va_start(ap, format);
+  retval = _vsnprintf_s(str, size, _TRUNCATE, format, ap);
+  va_end(ap);
+  return retval;
+}
+#define strdup _strdup
 #if defined(_WIN64)
 #define L2A(X) ((void *)(X))
 #define A2L(X) ((jlong)(X))
 #else
 #define L2A(X) ((void *)(unsigned long)(X))
 #define A2L(X) ((jlong)(unsigned long)(X))
-#endif
-#define snprintf sprintf_s
-#define strdup _strdup
-#else
-#if defined(_WIN32_WCE)
-#define snprintf _snprintf
 #endif
 #endif